Mock data added

+++ training/endorsements/helpers.py
@@ -1,6 +1,7 @@
 import requests
 from cachetools import cached, TTLCache
 from dotenv import load_dotenv
+from django.conf import settings
 from training.eud_header import eud_header
 
 load_dotenv()
@@ -8,6 +9,27 @@ load_dotenv()
 
 @cached(cache=TTLCache(maxsize=1024, ttl=60 * 10))
 def get_tier1_endorsements():
+    if settings.USE_CORE_MOCK:
+        return [
+            {
+                "id": 1,
+                "user_cid": 1601613,
+                "instructor_cid": 1439797,
+                "position": "EDDL_TWR",
+                "facility": 9,
+                "created_at": "2025-04-19T12:02:38.000000Z",
+                "updated_at": "2025-04-19T12:02:38.000000Z",
+            },
+            {
+                "id": 2,
+                "user_cid": 1601613,
+                "instructor_cid": 1439797,
+                "position": "EDDL_APP",
+                "facility": 9,
+                "created_at": "2025-04-19T12:02:38.000000Z",
+                "updated_at": "2025-04-19T12:02:38.000000Z",
+            },
+        ]
     endorsements = requests.get(
         "https://core.vateud.net/api/facility/endorsements/tier-1", headers=eud_header
     ).json()["data"]
@@ -15,6 +37,28 @@ def get_tier1_endorsements():
 
 
 def get_tier2_endorsements():
+    if settings.USE_CORE_MOCK:
+        return [
+            {
+                "id": 25,
+                "user_cid": 1439797,
+                "instructor_cid": 1439797,
+                "position": "EDXX_AFIS",
+                "facility": 9,
+                "created_at": "2024-02-29T22:39:33.000000Z",
+                "updated_at": "2024-02-29T22:39:33.000000Z",
+            },
+            {
+                "id": 26,
+                "user_cid": 1000000,
+                "instructor_cid": 1439797,
+                "position": "EDXX_AFIS",
+                "facility": 9,
+                "created_at": "2024-02-29T22:39:33.000000Z",
+                "updated_at": "2024-02-29T22:39:33.000000Z",
+            },
+        ]
+
     endorsements = requests.get(
         "https://core.vateud.net/api/facility/endorsements/tier-2", headers=eud_header
     ).json()["data"]
@@ -22,6 +66,8 @@ def get_tier2_endorsements():
 
 
 def remove_roster_and_endorsements(vatsim_id: int):
+    if settings.USE_CORE_MOCK:
+        return
     # Roster
     requests.delete(
         f"https://core.vateud.net/api/facility/roster/{vatsim_id}",
